+++ build.gradle
@@ -87,7 +87,9 @@ dependencies {
 
     // kakaoTolk 확인 테스트 용
     testImplementation 'com.github.database-rider:rider-junit5:1.32.0'
-}
+
+    // 욕 필터링 라이브러리
+    implementation 'io.github.vaneproject:badwordfiltering:1.0.0'}
 
 // chat gpt
 dependencyManagement {
