+++ src/main/java/com/undefinedus/backend/dto/request/discussion/DiscussionRegisterRequestDTO.java
@@ -1,5 +1,6 @@
 package com.undefinedus.backend.dto.request.discussion;
 
+import com.undefinedus.backend.global.validation.annotation.NoProfanity;
 import java.time.LocalDateTime;
 import lombok.Builder;
 import lombok.Data;
@@ -12,8 +13,10 @@ public class DiscussionRegisterRequestDTO {
 
     private Long member;  // 작성자
 
+    @NoProfanity
     private String title;    // 토론 제목
 
+    @NoProfanity
     private String content;      // 토론 주제 글
 
     private int agree; // 토론 찬성 반대 참여자 수를 세기 위해 필요 예(찬성 2)
