+++ src/main/java/com/undefinedus/backend/dto/request/discussionComment/DiscussionCommentRequestDTO.java
@@ -1,5 +1,6 @@
 package com.undefinedus.backend.dto.request.discussionComment;
 
+import com.undefinedus.backend.global.validation.annotation.NoProfanity;
 import lombok.Builder;
 import lombok.Data;
 
@@ -9,6 +10,7 @@ public class DiscussionCommentRequestDTO {
 
     private String voteType;  // 찬성/반대 의견
 
+    @NoProfanity
     private String content;  // 댓글 내용
 
 }
